chore: changing type of server_name
