+++ webserver/Makefile
@@ -10,7 +10,7 @@ CXXDEBUG = -D DEBUG=true
 SRCS =			main.cpp $(WEBSERV) $(SERVER) $(REQUEST) $(LOGGER) $(RESPONSE) \
 				$(UTILS)
 WEBSERV =		WebServ.cpp
-SERVER =		Server.cpp ServerParser.cpp ServerBuilder.cpp
+SERVER =		Server.cpp ServerParser.cpp ServerBuilder.cpp ServerGetters.cpp
 REQUEST =		Request.cpp RequestBuilder.cpp RequestParser.cpp
 RESPONSE =		Response.cpp ResponseBuilder.cpp StatusMessage.cpp HandleCGI.cpp
 LOGGER =		Logger.cpp
